Refactor: move miscs to statistics
