+++ tests/test_misc.py
@@ -21,45 +21,3 @@ def test_richset_iter() -> None:
             assert r == Something(1, "one")
         elif i == 1:
             assert r == Something(2, "two")
-
-
-def test_richset_is_empty() -> None:
-    assert RichSet.from_list([]).is_empty()
-    assert not RichSet.from_list([Something(1, "one")]).is_empty()
-
-
-def test_richset_is_non_empty() -> None:
-    assert RichSet.from_list([]).is_non_empty() is False
-    assert RichSet.from_list([Something(1, "one")]).is_non_empty() is True
-
-
-def test_richset_size() -> None:
-    assert RichSet.from_list([]).size() == 0
-    assert RichSet.from_list([Something(1, "one")]).size() == 1
-    assert (
-        RichSet.from_list(
-            [
-                Something(1, "one"),
-                Something(2, "two"),
-            ]
-        ).size()
-        == 2
-    )
-
-
-def test_richset_count() -> None:
-    assert RichSet.from_list([]).count(lambda r: True) == 0
-    assert RichSet.from_list([]).count(lambda r: False) == 0
-    rs = RichSet.from_list([Something(1, "one"), Something(2, "two")])
-    assert rs.count(lambda r: r.id == 1) == 1
-    assert rs.count(lambda r: r.id > 0) == 2
-    assert (
-        RichSet.from_list(
-            [
-                Something(1, "one"),
-                Something(2, "two"),
-                Something(3, "three"),
-            ]
-        ).count(lambda r: r.id % 2 == 0)
-        == 1
-    )
